Update main.py
Made it so a few more Kanji are included in the normal kanji set. User can now decide if they want translations. Fixed bug that caused links to go to the wrong files, causing self linking and liking to the word of a kanji instead of the kanji. Changed order of execution to probably make the links more correct.